crypto: msm: qcrypto: Close window during suspending for SMP
In the normal case, when a crypto operation is complete, driver process a
new request if there are any outstanding requests. In SMP, the suspending
may arrives in the window after driver completes the outstanding request,
and before driver starts a new operation. To close this window, a check
is added to make sure device is in BUS_HAS_BANDWIDTH to perform the next
operation.

Change-Id: Iccf73219dc55a5ad2af1f1d3a21f9a389db3d370
Signed-off-by: AnilKumar Chimata <anilc@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>